feat(cross-tests): fix several  issues
