+++ formulaire/frontend/src/components/ImagePickerMediaLibrary/index.tsx
@@ -74,7 +74,6 @@ export const ImagePickerMediaLibrary: FC<IImagePickerMediaLibraryProps> = ({
   };
 
   const handleMediaLibrarySuccess = (result: MediaLibraryResult) => {
-
     const mediaItem: MediaLibraryResult = Array.isArray(result) ? result[0] : result;
     const fileId: string = mediaItem._id ?? "";
