+++ formulaire/frontend/src/components/SentForm/index.tsx
@@ -3,10 +3,11 @@ import { ResourceCard } from "@cgi-learning-hub/ui";
 import { LOGO_PATH } from "~/core/constants";
 import { ISentFormProps } from "./types";
 import { useFormItemsIcons } from "~/hook/useFormItemsIcons";
+import { useMapToasterButtons } from "~/containers/HomeView/useMapToasterButtons";
 
 export const SentForm: FC<ISentFormProps> = ({ form, distributions, isSelected, handleSelect }) => {
   const { getSentFormPropertyItems } = useFormItemsIcons();
-
+  const { openFormResponseAction } = useMapToasterButtons();
   return (
     <ResourceCard
       key={form.id}
@@ -17,6 +18,9 @@ export const SentForm: FC<ISentFormProps> = ({ form, distributions, isSelected, 
       isSelected={isSelected(form)}
       onSelect={() => {
         handleSelect(form);
+      }}
+      onClick={() => {
+        void openFormResponseAction();
       }}
       propertyItems={getSentFormPropertyItems(form, distributions)}
       hasNoButtonOnFocus
