+++ formulaire/frontend/src/containers/HomeMainFormsTable/style.ts
@@ -1,0 +1,8 @@
+import { SxProps } from "@mui/material";
+import { centerBoxStyle } from "~/core/style/boxStyles";
+
+export const tablePaginationStyle = {
+  overflow: "unset",
+};
+
+export const iconBoxStyle: SxProps = { ...centerBoxStyle, gap: ".5rem" };
