+++ formulaire/frontend/src/containers/HomeMainSentFormTable/index.tsx
@@ -17,8 +17,8 @@ import { useSentFormColumns } from "./utils";
 import { Typography } from "@cgi-learning-hub/ui";
 import { TypographyVariant } from "~/core/style/themeProps";
 import { IForm } from "~/core/models/form/types";
-import { getPageForms, initialTableProps } from "../HomeMainTable/utils";
-import { tablePaginationStyle } from "../HomeMainTable/style";
+import { getPageForms, initialTableProps } from "../HomeMainFormsTable/utils";
+import { tablePaginationStyle } from "../HomeMainFormsTable/style";
 import { getFormDistributions, getFormStatusText, isFormFilled } from "~/core/models/form/utils";
 import { useFormatDateWithTime } from "~/hook/useFormatDateWithTime";
 import { ERROR_MAIN_COLOR, SUCCESS_MAIN_COLOR } from "~/core/style/colors";
