--- formulaire/frontend/src/containers/HomeMainTable/style.ts
@@ -1,3 +1,0 @@
-export const tablePaginationStyle = {
-  overflow: "unset",
-};
