+++ formulaire/frontend/src/containers/HomeView/useMapToasterButtons.ts
@@ -349,5 +349,5 @@ export const useMapToasterButtons = () => {
     selectedSentForm,
   ]);
 
-  return { leftButtons, rightButtons };
+  return { leftButtons, rightButtons, openFormResponseAction };
 };
